Fix/fee type (#6)
* convert response fee type to float

+++ src/near_omni_client/adapters/cctp/fee_service_types.py
@@ -5,7 +5,7 @@ class Fee(BaseModel):
     """Model representing a fee in the fee service."""
 
     finalityThreshold: int
-    minimumFee: int
+    minimumFee: float
 
 
 class GetFeeResponse(RootModel[list[Fee]]):
